Added a background processing setup to lot hosting. This will allow it to perform time expensive operations outside of the simulation thread.

+++ TSOClient/FSO.Server/Servers/Lot/Handlers/LotNegotiationHandler.cs
@@ -44,7 +44,7 @@ namespace FSO.Server.Servers.Lot.Handlers
                 return;
             }
 
-            if(Lots.TryAcceptClaim(request.EntityId, request.ClaimId, request.FromOwner))
+            if(Lots.TryAcceptClaim((int)request.EntityId, request.ClaimId, request.FromOwner))
             {
                 session.Write(new TransferClaimResponse
                 {
